更新Domain Panel中JTextArea的右键菜单
